Inline cmd/bb_portal/portal-service.go
