+++ cmd/bb_portal/main.go
@@ -4,9 +4,15 @@ import (
 	"context"
 	"flag"
 	"log/slog"
+	"net/http"
+	"net/http/httputil"
+	"net/url"
 	"os"
 	"time"
 
+	"entgo.io/contrib/entgql"
+	"github.com/99designs/gqlgen/graphql/handler"
+	"github.com/99designs/gqlgen/graphql/playground"
 	"github.com/fsnotify/fsnotify"
 	"github.com/gorilla/mux"
 	_ "github.com/mattn/go-sqlite3"
@@ -15,7 +21,9 @@ import (
 
 	"github.com/buildbarn/bb-portal/ent/gen/ent"
 	"github.com/buildbarn/bb-portal/ent/gen/ent/migrate"
+	"github.com/buildbarn/bb-portal/internal/api"
 	"github.com/buildbarn/bb-portal/internal/api/grpc/bes"
+	"github.com/buildbarn/bb-portal/internal/graphql"
 	"github.com/buildbarn/bb-portal/pkg/processing"
 	"github.com/buildbarn/bb-portal/pkg/proto/configuration/bb_portal"
 	"github.com/buildbarn/bb-storage/pkg/global"
@@ -153,4 +161,22 @@ func fatal(msg string, args ...any) {
 	// Workaround: No slog.Fatal.
 	slog.Error(msg, args...)
 	os.Exit(1)
+}
+
+func newPortalService(archiver processing.BlobMultiArchiver, dbClient *ent.Client, router *mux.Router) {
+	srv := handler.NewDefaultServer(graphql.NewSchema(dbClient))
+	srv.Use(entgql.Transactioner{TxOpener: dbClient})
+
+	router.PathPrefix("/graphql").Handler(srv)
+	router.Handle("/graphiql", playground.Handler("GraphQL Playground", "/graphql"))
+	router.Handle("/api/v1/bep/upload", api.NewBEPUploadHandler(dbClient, archiver)).Methods("POST")
+	router.PathPrefix("/").Handler(frontendServer())
+}
+
+func frontendServer() http.Handler {
+	targetURL := &url.URL{
+		Scheme: "http",
+		Host:   "localhost:3000",
+	}
+	return httputil.NewSingleHostReverseProxy(targetURL)
 }
